TF-500 - Update to move service logging configuration to a separate class
